+++ pkg/ucp/config.go
@@ -107,8 +107,10 @@ type RoutingConfig struct {
 // handles planes, and we need to support other kinds of resources.
 type InitializationConfig struct {
 	// Planes is a list of planes to create at startup.
-	Planes            []Plane `yaml:"planes,omitempty"`
-	ManifestDirectory string  `yaml:"manifestDirectory"`
+	Planes []Plane `yaml:"planes,omitempty"`
+
+	// ManifestDirectory is the directory which contains manifests.
+	ManifestDirectory string `yaml:"manifestDirectory"`
 }
 
 // Plane is a configuration entry for a plane resource. This is used to create a plane resource at startup.
@@ -126,6 +128,7 @@ type Plane struct {
 	Properties PlaneProperties `json:"properties" yaml:"properties"`
 }
 
+// PlaneProperties represents the properties of a plane resource.
 type PlaneProperties struct {
 	// ResourceProviders is a map of resource provider namespaces to their respective addresses.
 	//
